refactor: split toolcatalog JSON view types into views.go
The catalog.go file grew past the architecture-check 500-line guardrail
when the new source-attribution types landed. Splitting along the
natural seam — compiled-in tag/tool specs and surface utilities stay in
catalog.go; JSON-tagged API/UI views and their builders move to views.go.

No behavior change.
